Added BUILDING.md
* Fixed some warnings.
* Renamed i686 to x86.
* Added scrolling and \r support to terminal.

+++ kernel/arch/x86/vga.c
@@ -9,5 +9,5 @@ uint8_t vga_entry_color(enum vga_color fg, enum vga_color bg) {
 }
 
 uint16_t vga_entry(unsigned char uc, uint8_t color) {
-	return (uint16_t) uc | (uint16_t) color << 8;
+	return (uint16_t)(uc | color << 8);
 }
